% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_design.R
\name{opt_design}
\alias{opt_design}
\title{Find optimal sample zize and progression criteria}
\usage{
opt_design(rho_0, rho_1, alpha_nom, beta_nom, gamma_nom, max_n = NULL)
}
\arguments{
\item{rho_0}{Null hypothesis.}

\item{rho_1}{Alternative hypothesis.}

\item{alpha_nom}{Nominal upper constraint on alpha.}

\item{beta_nom}{Nominal upper constraint on beta.}

\item{gamma_nom}{Nominal upper constraint on gamma.}

\item{max_n}{Optional upper limit to use in search over sample sizes.}
}
\value{
A numeric vector containing the sample size, lower decision threshold,
and upper decision threshold; or NULL when no valid designs exist.
}
\description{
Find optimal sample zize and progression criteria
}
\examples{
rho_0 <- 0.5
rho_1 <- 0.7
alpha_nom <- 0.05
beta_nom <- 0.2
gamma_nom <- 0.5

opt_design(rho_0, rho_1, alpha_nom, beta_nom, gamma_nom)

}
