% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_pc.R
\name{opt_pc}
\alias{opt_pc}
\title{Find optimal procgression criteria for a fixed sample size}
\usage{
opt_pc(n, rho_0, rho_1, alpha_nom, beta_nom, gamma_nom)
}
\arguments{
\item{n}{Sample size.}

\item{rho_0}{Null hypothesis.}

\item{rho_1}{Alternative hypothesis.}

\item{alpha_nom}{Nominal upper constraint on alpha.}

\item{beta_nom}{Nominal upper constraint on beta.}

\item{gamma_nom}{Nominal upper constraint on gamma.}
}
\value{
A numeric vector containing the sample size, lower decision threshold,
and upper decision threshold.
}
\description{
Find optimal procgression criteria for a fixed sample size
}
\examples{
n <- 100
rho_0 <- 0.5
rho_1 <- 0.7
alpha_nom <- 0.05
beta_nom <- 0.2
gamma_nom <- 0.9

opt_pc(n, rho_0, rho_1, alpha_nom, beta_nom, gamma_nom)
}
