% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_pc_adjust_bin.R
\name{opt_pc_adjust_bin}
\alias{opt_pc_adjust_bin}
\title{Find optimal progression criteria for a fixed sample size and binary
outcome when an interval adjustment effect is known}
\usage{
opt_pc_adjust_bin(n, rho_0, rho_1, alpha_nom, beta_nom, tau_min, tau_max)
}
\arguments{
\item{n}{sample size.}

\item{rho_0}{null hypothesis.}

\item{rho_1}{alternative hypothesis.}

\item{alpha_nom}{nominal upper constraint on alpha.}

\item{beta_nom}{nominal upper constraint on beta.}

\item{tau_min}{lower limit of adjustment effect.}

\item{tau_max}{Upper limit if adjustment effect.}
}
\value{
A numeric vector containing the sample size, lower decision threshold,
and upper decision threshold (or NA when no valid designs exist), and
operating characteristics alpha, beta. Decision thresholds are on the outcome
scale.
}
\description{
Given a fixed sample size and a null and alternative hypothesis, this function
will determine upper and lower progression criteria such that alpha is controlled,
beta is minimised, and the width of the adjustment region is maximised.
Note that alpha and beta are defined taking into account an effect of adjustment
which will be in the given range.
}
\examples{
n <- 70
rho_0 <- 0.3
rho_1 <- 0.5
alpha_nom <- 0.05
beta_nom <- 0.2
tau_min <- 0.05
tau_max <- 0.1

opt_pc_adjust_bin(n, rho_0, rho_1, alpha_nom, beta_nom, tau_min, tau_max)

}
