% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_pc_adjust.R
\name{opt_pc_adjust_cont}
\alias{opt_pc_adjust_cont}
\title{Find optimal progression criteria for a fixed sample size and continuous
outcome when an interval adjustment effect is known}
\usage{
opt_pc_adjust_cont(
  n,
  rho_0,
  rho_1,
  alpha_nom,
  beta_nom,
  sigma,
  tau_min,
  tau_max
)
}
\arguments{
\item{n}{Sample size.}

\item{rho_0}{Null hypothesis.}

\item{rho_1}{Alternative hypothesis.}

\item{alpha_nom}{Nominal upper constraint on alpha.}

\item{beta_nom}{Nominal upper constraint on beta.}

\item{sigma}{Standard deviation of outcome (in the continuous case).}

\item{tau_min}{}

\item{tau_max}{}
}
\value{
A numeric vector containing the sample size, lower decision threshold,
and upper decision threshold (or NA when no valid designs exist), and
operating characteristics alpha, beta. Decision thresholds are on the outcome
z-statistic scale.
}
\description{
Given a fixed sample size and a null and alternative hypothesis, this function
will determine upper and lower progression criteria such that alpha is controlled
exactly, beta is minimised, and the width of the adjustment region is maximised.
Note that alpha and beta are defined taking into account an effect of adjustment
which will be in the given range.
}
\examples{
n <- 160
rho_0 <- 0
rho_1 <- 0.3
alpha_nom <- 0.05
beta_nom <- 0.2
sigma <- 1
tau_min <- 0.05
tau_max <- 0.1

opt_pc_adjust_cont(n, rho_0, rho_1, alpha_nom, beta_nom, sigma, tau_min, tau_max)
}
