% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_pc_cont.R
\name{opt_pc_cont}
\alias{opt_pc_cont}
\title{Find optimal progression criteria for a fixed sample size and continuous outcome}
\usage{
opt_pc_cont(
  n,
  rho_0,
  rho_1,
  sigma,
  alpha_nom,
  beta_nom,
  tau = c(0, 0),
  eta_0 = 0.5,
  eta_1 = NULL
)
}
\arguments{
\item{n}{sample size.}

\item{rho_0}{null hypothesis.}

\item{rho_1}{alternative hypothesis.}

\item{sigma}{standard deviation of the outcome.}

\item{alpha_nom}{nominal upper constraint on alpha.}

\item{beta_nom}{nominal upper constraint on beta.}

\item{tau}{vector with lower and upper bounds of adjustment effect.}

\item{eta_0}{probability of an incorrect go after an intermediate result.
Defaults to 0.5.}

\item{eta_1}{probability of an incorrect stop after an intermediate result.
Defaults to eta_0.}
}
\value{
A numeric vector containing the sample size, lower decision threshold,
and upper decision threshold (or NA when no valid designs exist), and
operating characteristics alpha, beta, gamma. Decision thresholds are on the
z-statistic scale.
}
\description{
Given a fixed sample size and a null and alternative hypothesis, this function
will determine upper and lower progression criteria such that alpha is controlled
exactly, beta is controlled or minimised when control is not feasible, and gamma
is minimised.
}
\examples{
n <- 200
rho_0 <- 0
rho_1 <- 0.3
sigma <- 1
alpha_nom <- 0.05
beta_nom <- 0.1

opt_pc_cont(n, rho_0, rho_1, sigma, alpha_nom, beta_nom)

}
