% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tout.R
\name{plot.tout}
\alias{plot.tout}
\title{Plot sampling distributions of three-outcome designs}
\usage{
\method{plot}{tout}(x, ...)
}
\arguments{
\item{x}{object of class \code{tout} as produced by \verb{tout_design().}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Takes an object of class \code{tout} and plots sampling distributions under the
null and alternative hypotheses, highlighting which portions correspond
to stop, pause, and go outcomes.
}
