% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tout.R
\name{print.tout}
\alias{print.tout}
\title{Print a tout object}
\usage{
\method{print}{tout}(x, ...)
}
\arguments{
\item{x}{object of class \code{tout} as produced by \code{tout_design()}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
no return value, called for side effects.
}
\description{
The default print method for a \code{tout} object.
}
