% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOut_design.R
\name{TOut_design_cont}
\alias{TOut_design_cont}
\title{Find optimal sample size and progression criteria}
\usage{
TOut_design_cont(
  rho_0,
  rho_1,
  sigma,
  alpha_nom,
  beta_nom,
  gamma_nom,
  eta = 0.5,
  tau = NULL,
  max_n = NULL
)
}
\arguments{
\item{rho_0}{null hypothesis.}

\item{rho_1}{alternative hypothesis.}

\item{sigma}{standard deviation of outcome.}

\item{alpha_nom}{nominal upper constraint on alpha.}

\item{beta_nom}{nominal upper constraint on beta.}

\item{gamma_nom}{nominal upper constraint on gamma.}

\item{eta}{probability of an incorrect decision under the null or alternative
after an intermediate result. Defaults to 0.5.}

\item{tau}{two element vector denoting lower and upper limits of the
effect of adjustment.}

\item{max_n}{optional upper limit to use in search over sample sizes.}
}
\value{
A numeric vector containing the sample size, lower decision threshold,
and upper decision threshold; or NULL when no valid designs exist.
}
\description{
Given a null and alternative hypothesis, this function finds the
lowest sample size such that a design with optimal progression criteria (as
determined by the function \code{opt_pc_cont}) satisfies upper constraints on three
operating characteristics.
}
\examples{
rho_0 <- 0
rho_1 <- 0.3
sigma <- 1
alpha_nom <- 0.05
beta_nom <- 0.2
gamma_nom <- 0.5

TOut_design_cont(rho_0, rho_1, sigma, alpha_nom, beta_nom, gamma_nom)

tau <- c(0.08, 0.12)

TOut_design_cont(rho_0, rho_1, sigma, alpha_nom, beta_nom, tau = tau)

}
