% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_design.R
\name{opt_design}
\alias{opt_design}
\title{Find optimal sample size and progression criteria}
\usage{
opt_design(
  rho_0,
  rho_1,
  alpha_nom,
  beta_nom,
  gamma_nom,
  max_n = NULL,
  n = NULL
)
}
\arguments{
\item{rho_0}{Null hypothesis.}

\item{rho_1}{Alternative hypothesis.}

\item{alpha_nom}{Nominal upper constraint on alpha.}

\item{beta_nom}{Nominal upper constraint on beta.}

\item{gamma_nom}{Nominal upper constraint on gamma.}

\item{max_n}{Optional upper limit to use in search over sample sizes.}

\item{n}{Optional sample size.}
}
\value{
A numeric vector containing the sample size, lower decision threshold,
and upper decision threshold; or NULL when no valid designs exist.
}
\description{
Given a null and alternative hypothesis, this function searches for the
lowest sample size such that a design with optimal progression criteria (as
determined by the function \code{opt_pc}) satisfies upper constraints on three
operating characteristics.
}
\examples{
rho_0 <- 0.5
rho_1 <- 0.7
alpha_nom <- 0.05
beta_nom <- 0.2
gamma_nom <- 0.5

opt_design(rho_0, rho_1, alpha_nom, beta_nom, gamma_nom)

}
