% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_pc.R
\name{opt_pc}
\alias{opt_pc}
\title{Find optimal progression criteria for a fixed sample size}
\usage{
opt_pc(
  n,
  rho_0,
  rho_1,
  alpha_nom,
  beta_nom,
  tau = c(0, 0),
  eta_0 = 0.5,
  eta_1 = NULL,
  x = NULL,
  sigma = NULL
)
}
\description{
Find optimal progression criteria for a fixed sample size
}
